feat (connect): added cache add method in ts
