+++ packages/connect/src/types.ts
@@ -1,3 +1,55 @@
+import { ListOptions, QueryOptions } from "./services/data";
+
+export enum Method {
+  GET = "GET",
+  POST = "POST",
+  PUT = "PUT",
+  DELETE = "DELETE",
+  PATCH = "PATCH",
+}
+
+enum Action {
+  QUERY = "_query",
+  BULK = "_bulk",
+  INDEX = "_index",
+}
+
+enum Service {
+  DATA = "data",
+  CACHE = "cache",
+}
+
+export interface Result {
+  ok: boolean;
+  id?: string;
+  msg?: string;
+}
+
+export interface Results<Type> {
+  ok: boolean;
+  docs: Type[];
+}
+
+export interface HyperData {
+  add: <Type>(body: Type) => Promise<Result>;
+  get: <Type>(id: string) => Promise<Type | Result>;
+  list: <T>(options?: ListOptions) => Promise<Results<T>>;
+  update: <Type>(id: string, doc: Type) => Promise<Result>;
+  remove: (id: string) => Promise<Result>;
+  query: <T>(selector: unknown, options?: QueryOptions) => Promise<Results<T>>;
+  index: (name: string, fields: string[]) => Promise<Result>;
+  bulk: <Type>(docs: Array<Type>) => Promise<Result>;
+}
+
+export interface HyperCache {
+  add: <Type>(key: string, value: Type, ttl?: string) => Promise<Result>;
+}
+
+export interface Hyper {
+  data: HyperData;
+  cache: HyperCache;
+}
+
 export interface HyperRequest {
   service: Service;
   method: Method;
